# Speed up width/height/depth by reducing copying