--- a/tests/module/mobject/mobject/test_mobject.py
+++ b/tests/module/mobject/mobject/test_mobject.py
@@ -150,7 +150,7 @@ def test_mobject_dimensions_has_points_and_children():
 
     # The width of a mobject should depend both on its points and
     # the points of all children mobjects.
-    assert outer_rect.width == 5  # 3 fom outer_rect, 2 from inner_rect
+    assert outer_rect.width == 5  # 3 from outer_rect, 2 from inner_rect
     assert outer_rect.height == 7  # 6 from outer_rect, 1 from inner_rect
     assert outer_rect.depth == 0