## Overview: What does this pull request change?
<!--changelog-start-->
The `width`, `height`, and `depth` attributes on Mobjects are calculated more quickly and consume less memory.
<!--changelog-end-->

I noticed large `Code` objects had some performance issues. For example, creating one with 59 lines of code and 1822 characters (based on a [real-world usage](https://www.youtube.com/watch?v=-aFtJt1_T3k)) took about 24.9 seconds, and calculating `height` (or `width` or `depth`) after creation took about 0.27s on my machine. Creating a profile and using SnakeViz showed that during the creation of Code, `height` calls took over 5 seconds of that time, due to a large amount of copying.
![Before profile, showing large amount of time calculating height ](https://user-images.githubusercontent.com/6819944/223767326-e742580d-250c-4c92-b3ce-4ce754b4e576.png)
![Zoomed in profile, showing height and width calculations spending a lot of time copying](https://user-images.githubusercontent.com/6819944/223767315-2b009c48-af7e-492f-ae18-8f4f2ddeedd7.png)

Note that after the "big" test in the profile, I also called `width` 10 times in a loop to try to gather more precise profiling/timing data.

## Motivation and Explanation: Why and how do your changes improve the library?
I changed the length calculations to not call `get_all_points` (which requires a lot of allocations, and copies of arrays), but just recursively find the largest value from a certain dimension. This reduced the length calls on my machine to about `0.077` seconds (3x speedup) and reduced the Code creation down to about 20.5 seconds. I plan to see about improving this further in a future PR.
![Profile showing Code taking 20.5 seconds](https://user-images.githubusercontent.com/6819944/223768736-38706d36-67f7-4656-ae0a-96ce0b9bd1f9.png)
![Profile showing height and width calculations are now dominated by calls to min/max](https://user-images.githubusercontent.com/6819944/223768733-4c80406d-ff5d-459d-926b-859fd45efdb0.png)

I also documented some of the (semi-internal) functions I explored and added tests for width/height/depth.

## Links to added or changed documentation pages
<!-- Please add links to the affected documentation pages (edit the description after opening the PR). The link to the documentation for your PR is https://manimce--####.org.readthedocs.build/en/####/, where #### represents the PR number. -->

<!-- Thank you again for contributing! Do not modify the lines below, they are for reviewers. -->
## Reviewer Checklist
- [ ] The PR title is descriptive enough for the changelog, and the PR is labeled correctly
- [ ] If applicable: newly added non-private functions and classes have a docstring including a short summary and a PARAMETERS section
- [ ] If applicable: newly added functions and classes are tested